INFRA-250 Eliminate "shebang" bug in Chord
Invoke Rscript rather than using the "shebang" (first line starting with
"!#") in the script to locate R. This eliminates the need to change the
Chord tooling for the pipeline.

Previously I had modified the tool in the bucket such that it would use
the `env` program in instead of an explicit `Rscript` path, which is how
we got to a working pipeline so far, but this change makes that one
redundant.