ACTIN-799 Add flag to rebuild Anaconda environment
Also try adding the Soundex library to Perl and make the on-VM imager
script version-aware so that it can propagate the version into the
tarball it creates for the Anaconda environment (perhaps premature but
if we are regularly updating it then compatibility is easily broken with
older versions).

Also make it so Anaconda is backed up to its own location in the bucket
instead of the R folder.