ACTIN-799 Tweak Anaconda cmds, push built tarball
The `conda` command was not working, it was consuming all memory on the
imaging VM and the machine would become unstable. Switching to `mamba`
instead immediately solved the problem. I also increased the memory on
the VM as it only had 4GB and it was quite close to that pretty much all
the time.

Also add some libraries that were needed to allow the install of the
Soundex library, update the library to the version actually installed by
`mamba`, and push a tarball of the completed Anaconda environment back
up to the `common-tools` bucket when complete.