ACTIN-799 Reorg files to copy to VM
It was getting a bit tedious naming them one at a time. I think some of
them can be deleted though.

Also make a copy of the `sv-prep` version `1.2.3` script from the
`common-tools` bucket so I can test whether my change to dump the "full
log" when something goes wrong to `stdout` works. I will contribute it
back to the bucket/project when we have confirmed.