ACTIN-799 Link the system Perl to Anaconda's one
This is a bit of a hack admittedly but the problem with the RepeatMasker
program is that it calls Perl in a "shebang" line (first line of script,
starting with '#!/usr/bin/perl') which means regardless of what is
configured to be the "default" Perl we're calling the system one
specifically.

Another way to deal with this would be to use `env` to locate the
configured Perl, eg `#!/usr/bin/env perl`, but that requires making yet
another change to the scripts in the bucket. Not quite sure this is the
right approach yet, going to mull (and see if it works in-pipeline...).