ACTIN-799 Fix another shebang line
The RepeatMasker script was calling the `famdb.py` script as an
executable rather than using its own Python3, I modified it to stop
doing that and am trying just with that file changed (ie the shebang
line in the `famdb.py` will not be relevant).

If that does not work I will also commit a changed `famdb.py`. The
tools.cmds have been modified to copy the RepeatMasker overtop the one
from the bucket which retains compatibility with other versions of the
pipeline that aren't using Debian 12 yet.

Also add the h5py dependency to Anaconda's configuration, this was
required by the `famdb.py` script.