ACTIN-799 Improve Anaconda automation (#562)
This was originally about adding the Perl Soundex module to the
Anaconda-managed environment but turned into something a little more
on the way. The motivation was to fix the failing virusbreakend stage
that was not noticed when testing the Debian 12 changes because
COLO829 does not have virus data.

Adding the Soundex module originally would have involved manually
extracting the Anaconda-managed environment on a VM and updating
it and then re-uploading the result to the common-tools bucket. This is
error prone and would cause problems if multiple versions were added
so I decided to change the imager to rebuild the Anaconda environment
without human involvement. Vindication for this approach came with
the later realisation that a Python script was missing the `h5py` library.

Once this was finished there were some other annoying things that came
up to do with the calling of the code that was skipped in the COLO829
runs. I have adjusted the scripts that were causing these issues in the
imager code itself to avoid confusion between versions.

Further to that last point I have created the `copy_to_imager_vm`
directory and everything under there is just copied to the imager
VM when it starts. Whichever parts of the script use files from there
can just reference them from their original names in `/tmp` on the VM.

The scripts that had to be adjusted are thus in the `copy_to_imager_vm`
directory. Changes were straightforward and just involved removing
"shebang" lines and improving logging. They can be differenced with
the original versions in the bucket (inspect the commands that copy
them overtop the originals in the `tools.cmd` file).

There was one other non-obvious change which was linking the system
Perl to the Anaconda one, this is not a beautiful solution but was
expedient. GRIDSS ecosystem is in the process of being deprecated
for the pipeline so I am fine with this as a short-term solution.

Co-authored-by: Ned Leitch <ned.leitch@gmail.com>