ACTIN-835 Restore AR commands for HMF JARs
This allows us to use JARs from the Artifact Registry transparently
instead of grabbing them from the bucket. Eventually this should replace
entirely the use of the bucket for internally-managed JARs.

Since the code to do this fetching was first written several other JARs
(Orange was first) have been added to the AR repository so use those
instead of the copies in the bucket too.

Also uncomment some code that looks like it was mistakenly disabled in
the private resource checkout script.