ACTIN-788 Go back to bucket-based JARs for now
There were differences between the JARs pulled in from Maven and the
ones in the `common-tools` bucket so for now I have reverted my change
that was using Artifact Registry for the JARs that were available in
there. We will push on that in the near future but for now don't rock
the boat by introducing binaries with unknown differences.

As an aside, the COLO run I did with the Artifact Regsitry JARs
completed successfully and I confirmed that all of the changes made in
ACTIN-835 to restore functionality lost through merge errors worked
properly, so when we again populate the results of
`VersionUtils.inArtifactRegistry()` with a non-empty list everything
should work as expected.