INFRA-696 Bump memory
This is starting to be a bit silly, but I can't quite work out how to
get the build to dump everything somewhere I can see it. Still have a
few more ideas but this may help.

Most recently the build failed with 134 which is an abort signal and
could have something to do with memory.