INFRA-696 Roll back unnecessary changes
These were all part of troubleshooting what was wrong with the pipeline5
build after the change to the Docker image, before I was able to
reproduce the problem locally.

The update of Mockito is required because it does not seem the
Debian-based image can still be built with Java 11 (without heroic
measures at least) so I just bit the bullet and updated the build to
Java 17.

I also removed the default location of `gcloud` for the `root` user as
it was no longer relevant with this Docker image.

I've left in the updates to the Surefire plugins and the small bit of
logging if the location of the `gcloud` SDK is defaulted as I think they
are useful changes.