DEV-4077: Improve handling of specific public resources version
When using a specific branch or commit from the
common-resources-public repository,
tag the commit used, and stop adding
"-unofficial" to the pipeline image family name. This last one was
mainly done such that such images can be used in the smoke
tests. These changes mirror ones already made in the master and
5.34-hotfix branches.